% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{adjacency}
\alias{adjacency}
\alias{connectivity}
\alias{connectivity}
\title{Returns the adjacency matrix}
\usage{
adjacency(x)

connectivity(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving an
adjacency matrix.}
}
\value{
the matrix of connections between clusters, inferred by the MST.

the matrix of connections between clusters, inferred by the MST.
}
\description{
Extract the adjacency matrix from a \code{SlingshotDataSet}.
}
\examples{
data("slingshotExample")
sds <- getLineages(rd, cl)
adjacency(sds)
}
