% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{connectivity}
\alias{connectivity}
\title{Returns the connectivity matrix}
\usage{
connectivity(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving a connectivity matrix.}
}
\value{
the matrix of connections between clusters.
}
