% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{curveControl}
\alias{curveControl}
\title{Returns the curve control parameters}
\usage{
curveControl(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving a set of
principal curves.}
}
\value{
the list of additional curve fitting parameters.
}
\description{
Extract the curve control parameters from a \code{SlingshotDataSet}.
}
\examples{
data("slingshotExample")
sds <- slingshot(rd, cl)
curveControl(sds)
}
