% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{curves}
\alias{curves}
\title{Returns the principal curves}
\usage{
curves(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving a set of
principal curves.}
}
\value{
the list of smooth lineage curves.
}
\description{
Extract the simultaneous principal curves from a \code{SlingshotDataSet}.
}
\examples{
data("slingshotExample")
sds <- slingshot(rd, cl)
curves(sds)
}
