% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot.R
\name{get_curves}
\alias{get_curves}
\title{Construct Smooth Curves for Each Lineage}
\usage{
get_curves(X, clus.labels, lineages, thresh = 1e-04, maxit = 100,
  stretch = 2, shrink = TRUE)
}
\arguments{
\item{X}{numeric, the \code{n x p} matrix of samples in a reduced dimensionality
space.}

\item{clus.labels}{character, a vector of length n denoting cluster labels.}

\item{lineages}{list, denotes which lineages each cluster is a part of and
contains the matrix defining the forest structure drawn on the clusters by
\code{\link{get_lineages}}.}

\item{thresh}{(optional) see documentation for \code{\link{principal.curve}}.}

\item{maxit}{(optional) see documentation for \code{\link{principal.curve}}.}

\item{stretch}{(optional) see documentation for \code{\link{principal.curve}}.}
}
\value{
A list of length \code{L}, equal to the number of lineages. Each element
  is an object of class \code{principal.curve} containing the following objects: 
  \itemize{ \item{\code{s}}{ a matrix of points along the curve corresponding to
  the projections of points in \code{X} onto the curve, ordered by pseudotime.}
  \item{\code{lambda}}{ a vector of pseudotime values in the same order as 
  \code{s}, representing each point's arclength along the curve.}
  \item{\code{dist}}{ the total squared distance between points used in the 
  construction of the curve and their projections onto the curve.}
  \item{\code{pseudotime}}{ a vector of pseudotime values of length \code{n},
  containing \code{NA} values for cells not represented by this lineage} }
}
\description{
This function takes a reduced data matrix \code{n x p}, a vector of
  cluster identities (optionally including -1's for "unclustered"), and a set of
  lineages consisting of paths through a forest constructed on the clusters. It
  constructs smooth curves for each lineage and returns the points along these
  curves corresponding to the orthogonal projections of each data point, along
  with corresponding arclength (\code{pseudotime} or \code{lambda}) values.
}
\details{
When there is only a single lineage, the curve-fitting algorithm is
  identical to that of \code{\link{principal.curve}}. When there are multiple
  lineages and \code{shrink=TRUE}, an additional step is added to the iterative 
  procedure, forcing curves to be similar in the neighborhood of shared points
  (ie., before they branch).
}
\examples{
data("slingshot_example")
lineages <- get_lineages(X, clus.labels, start.clus = 'a')
curves <- get_curves(X, clus.labels, lineages)
plot_curves(X, clus.labels, curves)

}

