% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot.R
\name{get_lineages}
\alias{get_lineages}
\title{Infer Lineage Structure from Clustered Samples}
\usage{
get_lineages(X, clus.labels, start.clus = NULL, end.clus = NULL,
  dist.fun = NULL, omega = Inf, distout = FALSE)
}
\arguments{
\item{X}{numeric, the \code{n x p} matrix of samples in a reduced dimensionality 
space.}

\item{clus.labels}{character, a vector of length n denoting cluster labels,
potentially including -1's for "unclustered."}

\item{start.clus}{(optional) character, indicates the cluster(s) *from* which
lineages will be drawn.}

\item{end.clus}{(optional) character, indicates the cluster(s) which will be
forced leaf nodes in their trees.}

\item{dist.fun}{(optional) function, method for calculating distances between
clusters. Must take two matrices as input, corresponding to points in
reduced-dimensional space. If the minimum cluster size is larger than the
number dimensions, the default is to use the joint covariance matrix to find
squared distance between cluster centers. If not, the default is to use the
diagonal of the joint covariance matrix.}

\item{omega}{(optional) numeric between 0 and 1 or Inf. This granularity
parameter determines the distance between every real cluster and the artificial
cluster, OMEGA. It is parameterized as a fraction of the largest distance
between two real clusters (hence, any value greater than 1 would result in a
single tree being returned and would be equivalent to setting omega = Inf, the
default)}

\item{distout}{(optional) logical, indicating whether the distance matrix between
the clusters should be included with output.}
}
\value{
a list with at least \code{L} items where \code{L} is the number of
  lineages identified. Each lineage is represented by a character vector with the
  names of the clusters included in that lineage. Additional items may include:
  \code{forest}, the connectivity matrix; \code{C}, a clusters x lineages
  identity matrix; \code{start.clus} and \code{end.clus}, the starting and ending
  cluster(s); \code{start.given} and \code{end.given}, logical values indicating
  whether the starting and ending clusters were specified a priori; \code{dist},
  the pairwise cluster distance matrix.
}
\description{
Given a reduced-dimension data matrix \code{n x p} and a vector of
cluster identities (potentially including -1's for "unclustered"), this function
infers a forest structure on the clusters and returns paths through the forest
that can be interpreted as lineages.
}
\details{
The connectivity matrix, denoted \code{forest}, is learned by fitting a
  (possibly constrained) minimum-spanning tree on the clusters and the artificial
  cluster, OMEGA, which is a distance \code{omega} from every real cluster.

Once the \code{forest} is known, lineages are identified in any tree
  with at least two clusters. For a given tree, if there is an annotated starting
  cluster, every possible path out of a starting cluster and ending in a leaf
  that isn't another starting cluster will be returned. If no starting cluster is
  annotated, every leaf will be considered as a potential starting cluster and
  whichever configuration produces the longest average lineage length (in terms
  of number of clusters included) will be returned.
}
\examples{
data("slingshot_example")
get_lineages(X, clus.labels)
get_lineages(X, clus.labels, start.clus = 'a')

}

