% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLineages.R
\docType{methods}
\name{getLineages,matrix,character-method}
\alias{getLineages,matrix,character-method}
\alias{getLineages}
\title{Infer Lineage Structure from Clustered Samples}
\usage{
\S4method{getLineages}{matrix,character}(reducedDim, clusLabels,
  start.clus = NULL, end.clus = NULL, dist.fun = NULL, omega = NULL)
}
\arguments{
\item{reducedDim}{numeric, the \code{n} by \code{p} matrix of samples in a reduced 
dimensionality space.}

\item{clusLabels}{character, a vector of length \code{n} denoting cluster labels,
optionally including \code{-1}'s for "unclustered." If \code{reducedDim} is a
\code{SlingshotDataSet}, cluster labels will be taken from it.}

\item{start.clus}{(optional) character, indicates the cluster(s) *from* which
lineages will be drawn.}

\item{end.clus}{(optional) character, indicates the cluster(s) which will be
forced leaf nodes in their trees.}

\item{dist.fun}{(optional) function, method for calculating distances between
clusters. Must take two matrices as input, corresponding to points in
reduced-dimensional space. If the minimum cluster size is larger than the
number dimensions, the default is to use the joint covariance matrix to find
squared distance between cluster centers. If not, the default is to use the
diagonal of the joint covariance matrix.}

\item{omega}{(optional) numeric between 0 and 1. This granularity
parameter determines the distance between every real cluster and the artificial
cluster, OMEGA. It is parameterized as a fraction of the largest distance
between two real clusters (hence, any value greater than 1 would result in a
single tree being returned and would be equivalent to setting \code{omega = Inf},
which is the default behavior).}
}
\value{
An object of class \code{\link{SlingshotDataSet}} containing the 
arguments provided to \code{getLineages} as well as the following new
elements:
\itemize{
\item{\code{lineages}}{ a list of \code{L} items, where \code{L} is the number 
  of lineages identified. Each lineage is represented by a character vector 
  with the names of the clusters included in that lineage, in order.}
\item{\code{connectivity}}{ the inferred cluster connectivity matrix.}
\item{\code{lineage.control$start.given},\code{lineage.control$end.given}}
  { logical values indicating whether the starting and ending clusters were 
  specified a priori.} 
\item{\code{lineage.control$dist}}{ the pairwise cluster distance matrix.}}
}
\description{
Given a reduced-dimension data matrix \code{n} by \code{p} and a vector of
cluster identities (potentially including -1's for "unclustered"), this function
infers a forest structure on the clusters and returns paths through the forest
that can be interpreted as lineages.
}
\details{
The \code{connectivity} matrix is learned by fitting a
  (possibly constrained) minimum-spanning tree on the clusters and the artificial
  cluster, OMEGA, which is a fixed distance away from every real cluster. This 
  effectively limits the maximum branch length in the MST to twice the chosen 
  distance, meaning that the output may contain multiple trees.

Once the \code{connectivity} is known, lineages are identified in any tree
  with at least two clusters. For a given tree, if there is an annotated starting
  cluster, every possible path out of a starting cluster and ending in a leaf
  that isn't another starting cluster will be returned. If no starting cluster is
  annotated, every leaf will be considered as a potential starting cluster and
  whichever configuration produces the longest average lineage length (in terms
  of number of clusters included) will be returned.
}
\examples{
data("slingshotExample")
sds <- getLineages(rd, cl, start.clus = '5')

plot(rd, col = cl, asp = 1)
lines(sds, type = 'l', lwd = 3)

}
