% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{lineageControl}
\alias{lineageControl}
\title{Returns the lineage control parameters}
\usage{
lineageControl(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving lineages.}
}
\value{
the list of additional lineage inference parameters.
}
