% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotting.R
\docType{methods}
\name{plotGenePseudotime}
\alias{plotGenePseudotime}
\alias{plotGenePseudotime,character,SlingshotDataSet,matrix-method}
\title{Plot Gene Expression by Pseudotime}
\usage{
plotGenePseudotime(gene, sds, exprs, ...)

\S4method{plotGenePseudotime}{character,SlingshotDataSet,matrix}(gene, sds,
  exprs, loess = TRUE, loessCI = TRUE, ...)
}
\arguments{
\item{gene}{character, the name of the gene to be plotted.}

\item{sds}{a \code{SlingshotDataSet} with curves fitted.}

\item{exprs}{matrix of expression values (genes by samples). Must include a 
row with a row name of \code{gene}.}

\item{...}{additional parameters to be passed to \code{\link{plot}}.}

\item{loess}{logical, whether to include a loess fit in each plot (default is
\code{TRUE}).}

\item{loessCI}{logical, whether to include a confidence band around the loess
curve (default is \code{TRUE}).}

\item{gene}{character, the name of the gene to be plotted}

\item{sds}{a SlingshotDataSet containing cell pseudotimes}

\item{exprs}{the genes-by-samples matrix of expression values.}
}
\value{
returns \code{NULL}.
}
\description{
Show the gene expression pattern for an individual gene along
lineages inferred by \code{\link{slingshot}}.
}
\examples{
data("slingshotExample")
sds <- slingshot(rd, cl, start.clus = "1")
ex <- matrix(c(rchisq(100,1),rchisq(20,3),rchisq(20,6)),nrow=1)
rownames(ex) <- 'Gene-1'
plotGenePseudotime('Gene-1', sds, ex)

}
