% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\docType{methods}
\name{pseudotime}
\alias{pseudotime}
\alias{curveWeights,SlingshotDataSet-method}
\title{Get Slingshot pseudotime values}
\usage{
\S4method{pseudotime}{SlingshotDataSet}(x, na = TRUE)

\S4method{curveWeights}{SlingshotDataSet}(x)
}
\arguments{
\item{x}{a \code{SlingshotDataSet} object.}

\item{na}{logical. If \code{TRUE} (default), cells that are not assigned to a
lineage will have a pseudotime value of \code{NA}. Otherwise, their
arclength along the curve will be returned.}
}
\description{
Extract the matrix of pseudotime values or cells' weights along
  each lineage.
}
