% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{reducedDim}
\alias{reducedDim}
\alias{reducedDim<-}
\title{Returns the reduced dimensional representation of a dataset.}
\usage{
reducedDim(x)

reducedDim(x) <- value
}
\arguments{
\item{x}{an object that describes a dataset or a model involving reduced
dimensional data.}
}
\value{
the matrix representing the reduced dimensional data.

Updated object with new reduced dimensional matrix.
}
\description{
Returns the reduced dimensional representation of a dataset.
}
\examples{
rd <- matrix(data=rnorm(100), ncol=2)
cl <- sample(letters[1:5], 50, replace = TRUE)
sds <- newSlingshotDataSet(rd, cl)
reducedDim(sds)
}
