% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\docType{methods}
\name{slingAdjacency}
\alias{slingAdjacency}
\alias{slingAdjacency,SlingshotDataSet-method}
\alias{slingAdjacency,SingleCellExperiment-method}
\title{Extract Slingshot adjacency matrix}
\usage{
slingAdjacency(x)

\S4method{slingAdjacency}{SlingshotDataSet}(x)

\S4method{slingAdjacency}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{an object containing \code{\link{slingshot}} output.}
}
\value{
the matrix of connections between clusters, inferred by the MST.
}
\description{
Extract the adjacency matrix from an object containing
  \code{\link{slingshot}} output.
}
\section{Methods (by class)}{
\itemize{
\item \code{SlingshotDataSet}: returns the adjacency matrix between
clusters from a \code{\link{SlingshotDataSet}} object.

\item \code{SingleCellExperiment}: returns the adjacency matrix between
clusters from a \code{\link{SingleCellExperiment}} object.
}}

\examples{
data("slingshotExample")
sds <- getLineages(rd, cl)
slingAdjacency(sds)
}
