% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchID.R
\name{slingBranchGraph}
\alias{slingBranchGraph}
\title{Construct graph of slingshot branch labels}
\usage{
slingBranchGraph(x, thresh = NULL)
}
\arguments{
\item{x}{an object containing slingshot output.}

\item{thresh}{the minimum weight of assignment required to assign a cell to a
lineage (default = 1/L)}
}
\description{
Builds a graph describing the relationships between the
  different branch assignments
}
