% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchID.R
\name{slingBranchID}
\alias{slingBranchID}
\title{Get slingshot branch labels}
\usage{
slingBranchID(x, thresh = NULL)
}
\arguments{
\item{x}{an object containing slingshot output.}

\item{thresh}{the minimum weight of assignment required to assign a cell to a
lineage (default = 1/L)}
}
\description{
Extracts a categorical variable describing which lineage (or
  combination of lineages) each cell is assigned to.
}
