% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\name{slingClusterLabels}
\alias{slingClusterLabels}
\alias{slingClusterLabels,SlingshotDataSet-method}
\alias{slingClusterLabels,SingleCellExperiment-method}
\title{Extract cluster labels used by Slingshot}
\usage{
slingClusterLabels(x)

\S4method{slingClusterLabels}{SlingshotDataSet}(x)

\S4method{slingClusterLabels}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{an object containing \code{\link{slingshot}} output.}
}
\value{
a vector of cluster labels or a matrix of cluster assignment weights.
}
\description{
Extract the cluster labels used by \code{\link{slingshot}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{SlingshotDataSet}: returns the cluster labels stored in a
\code{\link{SlingshotDataSet}} object.

\item \code{SingleCellExperiment}: returns the cluster labels used by
\code{\link{slingshot}} in a \code{\link{SingleCellExperiment}} object.
}}

\examples{
data("slingshotExample")
sds <- getLineages(rd, cl)
slingClusterLabels(sds)
}
