% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{slingParams}
\alias{slingParams}
\title{Methods for parameters used by Slingshot}
\usage{
slingParams(x)
}
\arguments{
\item{x}{an object that includes output from Slingshot, either a 
\code{SlingshotDataSet} or \code{SingleCellExperiment} object.}
}
\value{
the list of additional parameters used by Slingshot.
}
\description{
Extracts additional control parameters used by Slingshot in 
lineage inference and fitting simultaneous principal curves.
}
\examples{
data("slingshotExample")
sds <- slingshot(rd, cl, start.clus = '5')
slingParams(sds)
}
