% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R, R/plotting.R
\docType{methods}
\name{show,SlingshotDataSet-method}
\alias{show,SlingshotDataSet-method}
\alias{reducedDim,SlingshotDataSet-method}
\alias{clus.labels,SlingshotDataSet-method}
\alias{connectivity,SlingshotDataSet-method}
\alias{lineage.control,SlingshotDataSet-method}
\alias{curves,SlingshotDataSet-method}
\alias{pseudotime,SlingshotDataSet-method}
\alias{weights,SlingshotDataSet-method}
\alias{curve.control,SlingshotDataSet-method}
\alias{reducedDim<-,SlingshotDataSet-method}
\alias{clus.labels<-,SlingshotDataSet-method}
\alias{lines,SlingshotDataSet-method}
\title{Methods for the \code{SlingshotDataSet} class}
\usage{
\S4method{show}{SlingshotDataSet}(object)

\S4method{reducedDim}{SlingshotDataSet}(x)

\S4method{clus.labels}{SlingshotDataSet}(x)

\S4method{connectivity}{SlingshotDataSet}(x)

\S4method{lineage.control}{SlingshotDataSet}(x)

\S4method{curves}{SlingshotDataSet}(x)

\S4method{pseudotime}{SlingshotDataSet}(x)

\S4method{weights}{SlingshotDataSet}(x)

\S4method{curve.control}{SlingshotDataSet}(x)

\S4method{reducedDim}{SlingshotDataSet}(x) <- value

\S4method{clus.labels}{SlingshotDataSet}(x) <- value

\S4method{lines}{SlingshotDataSet}(x, type = NULL, dims = 1:2, ...)
}
\description{
Helper functions for manipulating objects of the \code{SlingshotDataSet}
class.
}
