% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{SlingshotDataSet}
\alias{SlingshotDataSet}
\alias{SlingshotDataSet,data.frame,ANY-method}
\alias{SlingshotDataSet,matrix,numeric-method}
\alias{SlingshotDataSet,matrix,factor-method}
\alias{SlingshotDataSet,matrix,character-method}
\title{Constructor function for the \code{SlingshotDataSet} class}
\usage{
SlingshotDataSet(reducedDim, clusterLabels, ...)

\S4method{SlingshotDataSet}{data.frame,ANY}(reducedDim, clusterLabels, ...)

\S4method{SlingshotDataSet}{matrix,numeric}(reducedDim, clusterLabels, ...)

\S4method{SlingshotDataSet}{matrix,factor}(reducedDim, clusterLabels, ...)

\S4method{SlingshotDataSet}{matrix,character}(reducedDim, clusterLabels,
  lineages = list(), connectivity = matrix(NA, 0, 0),
  lineageControl = list(), curves = list(), pseudotime = matrix(NA, 0, 0),
  curveWeights = matrix(NA, 0, 0), curveControl = list())
}
\arguments{
\item{reducedDim}{matrix. An \code{n} by \code{p} numeric matrix or data frame
giving the coordinates of the cells in a reduced dimensionality space.}

\item{clusterLabels}{character. A character vector of length \code{n} denoting
each cell's cluster label.}

\item{...}{Additional elements to specify, see
\code{\link{SlingshotDataSet-class}}.}
}
\value{
A \code{SlingshotDataSet} object.
}
\description{
The \code{SlingshotDataSet} constructor creates an object of the
 class \code{SlingshotDataSet}. Objects of this type will also be returned by
 \code{\link{getLineages}} and \code{\link{getCurves}}.

Additional functions extract or replace elements of a
 \code{SlingshotDataSet}.
}
\examples{

rd <- matrix(data=rnorm(200), ncol=2)
cl <- sample(letters[1:5], 100, replace = TRUE)
sds <- SlingshotDataSet(rd, cl)

}
\seealso{
\code{\link{SlingshotDataSet-class}}
}
