% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\docType{data}
\name{slingshotExample}
\alias{slingshotExample}
\alias{rd}
\alias{cl}
\alias{cl}
\title{Bifurcating lineages data}
\format{A matrix of coordinates in two dimensions, representing $140$ cells,
  and a numeric vector $140$ corresponding cluster labels.}
\source{
Simulated data provided with the \code{slingshot} package.
}
\usage{
rd

cl
}
\description{
A simulated low-dimensional representation of two bifurcating
  lineages (\code{rd}) and a vector of cluster labels generated by $k$-means
  with $K = 5$ (\code{cl}).
}
\keyword{datasets}
