% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\docType{data}
\name{splatterParameters}
\alias{splatterParameters}
\alias{params}
\title{Parameters for simulating a single lineage}
\format{A \code{\link[splatter:SplatParams]{SplatParams}} object with some
  path-related parameters set manually.}
\source{
Most parameters were learned by
  \code{\link[splatter:splatEstimate]{splatEstimate}} and based on the
  \code{\link[HSMMSingleCell:HSMM]{HSMM}} dataset.
}
\usage{
params
}
\description{
A \code{SplatParams} object containing the parameters
  needed to perform the simulation shown in the Slingshot vignette. Values
  are designed to produce a single, non-branching trajectory. Code for
  producing these parameters can be found in the Slingshot vignette.
}
\keyword{datasets}
