% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\docType{methods}
\name{clusterLabels}
\alias{clusterLabels}
\alias{clusterLabels<-}
\alias{clusterLabels,SlingshotDataSet-method}
\alias{clusterLabels,SingleCellExperiment-method}
\title{Returns the cluster labels}
\usage{
clusterLabels(x)

clusterLabels(x) <- value

\S4method{clusterLabels}{SlingshotDataSet}(x)

\S4method{clusterLabels}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving cluster
labels.}

\item{x}{an object that describes a dataset or a model involving cluster
labels.}
}
\value{
the vector of cluster labels.

Updated object with new vector of cluster labels.

A matrix of cluster weights for each cell or a vector of cluster
  assignments.
}
\description{
Extract cluster labels from a \code{SlingshotDataSet}.

Extract cluster labels, either a character vector or matrix of 
  weights.
}
\section{Methods (by class)}{
\itemize{
\item \code{SlingshotDataSet}: extracts cluster labels from a
\code{SlingshotDataSet} object.

\item \code{SingleCellExperiment}: extracts the cluster labels used by Slingshot from 
a \code{\link{SingleCellExperiment}} object.
}}

\examples{
rd <- matrix(data=rnorm(100), ncol=2)
cl <- sample(letters[1:5], 50, replace = TRUE)
sds <- newSlingshotDataSet(rd, cl)
clusterLabels(sds)
rd <- matrix(data=rnorm(100), ncol=2)
cl <- sample(letters[1:5], 50, replace = TRUE)
sds <- newSlingshotDataSet(rd, cl)
clusterLabels(sds)
}
