% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{clusterLabels}
\alias{clusterLabels}
\alias{clusterLabels<-}
\title{Returns the cluster labels}
\usage{
clusterLabels(x)

clusterLabels(x) <- value
}
\arguments{
\item{x}{an object that describes a dataset or a model involving cluster
labels.}
}
\value{
the vector of cluster labels.

Updated object with new vector of cluster labels.
}
\description{
Extract cluster labels from a \code{SlingshotDataSet}.
}
\examples{
rd <- matrix(data=rnorm(100), ncol=2)
cl <- sample(letters[1:5], 50, replace = TRUE)
sds <- newSlingshotDataSet(rd, cl)
clusterLabels(sds)
}
