% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot2.R
\name{get_curves}
\alias{get_curves}
\title{Construct Smooth Curves for Each Lineage}
\usage{
get_curves(X, clus.labels, lineages, thresh = 1e-04, maxit = 100,
  stretch = 2, trace = FALSE, shrink = TRUE)
}
\arguments{
\item{X}{numeric, the nxp matrix of samples in a reduced dimensionality space}

\item{clus.labels}{character, a vector of length n denoting cluster labels}

\item{lineages}{list, denotes which lineages each cluster is a part of and contains the matrix defining the forest structure drawn on the clusters by \code{get_lineages}.}

\item{thresh}{(optional)}

\item{maxit}{(optional)}

\item{stretch}{(optional)}

\item{trace}{(optional)}
}
\value{
TODO
}
\description{
This function takes a reduced data matrix nxp, a vector of cluster identities (optionally including -1's for "unclustered"), and a set of lineages consisting of paths through a forest constructed on the clusters. It constructs smooth curves for each lineage and returns the points along these curves corresponding to the orthogonal projections of each data point, along with lambda (pseudotime) values.
}
\details{
TODO
}
\examples{
data("toy_data")
lineages <- get_lineages(X, clus.labels, start.clus = 'a')
get_curves(X, clus.labels, lineages)

}

