% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{lineages}
\alias{lineages}
\title{Returns the lineages}
\usage{
lineages(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving lineages}
}
\value{
the list of lineages, represented by ordered sets of clusters.
}
