% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pairs-SlingshotDataSet}
\alias{pairs-SlingshotDataSet}
\alias{pairs.SlingshotDataSet}
\title{Pairs plot of Slingshot output}
\usage{
\method{pairs}{SlingshotDataSet}(x, labels, col = NULL, cex = 1, lwd = 2,
  ..., horInd = 1:nc, verInd = 1:nc, lower.panel = FALSE,
  upper.panel = TRUE, diag.panel = NULL, text.panel = textPanel,
  label.pos = 0.5 + has.diag/3, line.main = 3, cex.labels = NULL,
  font.labels = 1, row1attop = TRUE, gap = 1, show.constraints = FALSE,
  type = NULL, pch = 16)
}
\arguments{
\item{x}{a \code{SlingshotDataSet} with results to be plotted.}

\item{...}{additional parameters for \code{plot} or \code{axis}, see 
\code{\link{pairs}}.}

\item{type}{character, the type of output to be plotted, can be one of 
\code{"lineages"}, \code{curves}, or \code{both} (by partial matching), see
Details for more.}
}
\value{
returns \code{NULL}.
}
\description{
A tool for quickly visualizing lineages inferred by
  \code{slingshot}.
}
\details{
#' @rdname SlingshotDataSet-plot3d
#' @export
setMethod(
  f = "lines3d",
  signature = "SlingshotDataSet",
  definition = function(x,
                        type = NULL,
                        dims = 1:3,
                        ...) {
    lines3d.SlingshotDataSet(x, type = type, add = TRUE, dims = dims, ...)
  }
)

If \code{type == 'lineages'}, straight line connectors between
  cluster centers will be plotted. If \code{type == 'curves'}, simultaneous
  princiapl curves will be plotted.
  

When \code{type} is not specified, the function will first check the
  \code{curves} slot and plot the curves, if present. Otherwise,
  \code{lineages} will be plotted, if present.
}
\examples{
data("slingshotExample")
sds <- slingshot(rd, cl, start.clus = "1")
pairs(sds, type = 'curves')

}
