% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_curves}
\alias{plot_curves}
\title{Plot smooth curves}
\usage{
plot_curves(X, clus.labels, curves, threeD = TRUE, dim = NA, col = NA)
}
\arguments{
\item{X}{numeric, the \code{n x p} matrix of samples in a reduced dimensionality space.}

\item{clus.labels}{character, a vector of length n denoting cluster labels.}

\item{curves}{list, the output of \code{\link{get_curves}}, this list includes matrices for each curve in the reduced-dimensional space.}

\item{threeD, }{logical indicates whether to make a 3D plot with the \code{rgl} package.}

\item{dim, }{total number of dimensions to be shown in a series of two-dimensional plots, similar to \code{pairs} plots (only applicable if \code{threeD} is false).}

\item{col, }{(optional) vector of colors to use for denoting clusters.}
}
\value{
returns \code{NULL}.
}
\description{
\code{plot_curves} visualizes the output from \code{\link{get_curves}}, showing cells (colored by cluster) in a reduced-dimensional space with smooth curves representing each inferred lineage.
}
\details{
TODO
}
\examples{
data("slingshot_example")
lineages <- get_lineages(X, clus.labels, start.clus = 'a')
get_curves(X, clus.labels, lineages)
plot_curves(X, clus.labels, curves, threeD = FALSE)

}

