% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotting.R
\docType{methods}
\name{plotGenePseudotime}
\alias{plotGenePseudotime}
\alias{plotGenePseudotime,SlingshotDataSet-method}
\alias{plotGenePseudotime,SingleCellExperiment-method}
\title{Plot Gene Expression by Pseudotime}
\usage{
plotGenePseudotime(data, ...)

\S4method{plotGenePseudotime}{SlingshotDataSet}(data, gene, exprs,
  loess = TRUE, loessCI = TRUE, ...)

\S4method{plotGenePseudotime}{SingleCellExperiment}(data, gene, exprs,
  loess = TRUE, loessCI = TRUE, ...)
}
\arguments{
\item{data}{an object containing \code{\link{slingshot}} output, either a
\code{\link{SlingshotDataSet}} or a \code{\link{SingleCellExperiment}}
object.}

\item{...}{additional parameters to be passed to \code{\link{plot}}.}

\item{gene}{the gene to be plotted. If \code{exprs} is provided, this may be
either the gene name or its row index in \code{exprs}. Otherwise, this is
assumed to be a vector of scaled expression values.}

\item{exprs}{the genes-by-samples matrix of scaled expression values (log
counts or normalized log counts).}

\item{loess}{logical, whether to include a loess fit in each plot (default is
\code{TRUE}).}

\item{loessCI}{logical, whether to include a confidence band around the loess
curve (default is \code{TRUE}).}
}
\value{
returns \code{NULL}.
}
\description{
Show the gene expression pattern for an individual gene along
lineages inferred by \code{\link{slingshot}}.
}
\examples{
data("slingshotExample")
sds <- slingshot(rd, cl, start.clus = "1")
ex <- matrix(c(rchisq(100,1),rchisq(20,3),rchisq(20,6)),nrow=1)
rownames(ex) <- 'Gene-1'
plotGenePseudotime(sds, 'Gene-1', ex)

}
