% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,SlingshotDataSet-method}
\alias{predict,SlingshotDataSet-method}
\title{Predict from a Slingshot model}
\usage{
\S4method{predict}{SlingshotDataSet}(object, newdata = NULL)
}
\arguments{
\item{object}{a \code{\link{SlingshotDataSet}} containing simultaneous 
principal curves to use for prediction.}

\item{newdata}{a matrix or data frame of new points in the same 
reduced-dimensional space as the original input to \code{slingshot} (or
\code{getLineages}).}
}
\value{
A \code{SlingshotDataSet} object based on the input \code{newdata}.
  New cells are treated as "unclustered" and the \code{lineages} and
  \code{adjacency} slots are intentionally left blank, to distinguish these
  results from the original \code{slingshot} output. The \code{curves} slot
  represents the projections of each new cell onto the existing curves. As
  with standard \code{slingshot} output, the lineage-specific pseudotimes and
  assignment weights can be accessed via the functions
  \code{\link{slingPseudotime}} and \code{\link{slingCurveWeights}}.
}
\description{
Map new observations onto simultaneous principal curves fitted
  by \code{slingshot}.
}
\details{
This function is a method for the generic function \code{predict}
  with \code{signature(object = "SlingshotDataSet")}. If no \code{newdata}
  argument is provided, it will return the original results, given by
  \code{object}.
}
\examples{
data("slingshotExample")
sds <- slingshot(rd, cl, start.clus = '1')

x <- cbind(runif(100, min = -5, max = 10), runif(100, min = -4, max = 4))
predict(sds, x)

}
\seealso{
\code{\link{slingshot}}, \code{\link{SlingshotDataSet}}
}
