% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{pseudotime}
\alias{pseudotime}
\alias{curveWeights}
\title{Get Slingshot pseudotime values}
\usage{
pseudotime(x, ...)

curveWeights(x)
}
\arguments{
\item{x}{a \code{SlingshotDataSet} object.}

\item{...}{additional parameters to be passed to object-specific methods.}
}
\value{
an \code{n} by \code{L} matrix representing each cell's pseudotime
  along each lineage.

an \code{n} by \code{L} matrix of cell weights along
  each lineage.
}
\description{
Extract the matrix of pseudotime values or cells' weights along
  each lineage.
}
\details{
#' @rdname clusterLabels 
#' @return Updated object with new vector of cluster labels.
#' @export
setGeneric(name = "clusterLabels<-", 
           signature = "x",
           def = function(x, value) standardGeneric("clusterLabels<-"))
}
\examples{
data("slingshotExample")
sds <- slingshot(rd, cl)
pseudotime(sds)
data("slingshotExample")
sds <- slingshot(rd, cl)
curveWeights(sds)
}
