% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\docType{methods}
\name{slingCurves}
\alias{slingCurves}
\alias{slingCurves,SlingshotDataSet-method}
\alias{slingCurves,SingleCellExperiment-method}
\title{Extract simultaneous principal curves}
\usage{
slingCurves(x)

\S4method{slingCurves}{SlingshotDataSet}(x)

\S4method{slingCurves}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{an object containing \code{\link{slingshot}} output.}
}
\value{
the list of smooth lineage curves, each of which is a
  \code{\link[princurve]{principal.curve}} object.
}
\description{
Extract the simultaneous principal curves from an object
  containing \code{\link{slingshot}} output.
}
\section{Methods (by class)}{
\itemize{
\item \code{SlingshotDataSet}: returns the list of smooth lineage curves from a
\code{\link{SlingshotDataSet}} object.

\item \code{SingleCellExperiment}: returns the list of smooth lineage curves from a
\code{\link{SingleCellExperiment}} object.
}}

\examples{
data("slingshotExample")
sds <- slingshot(rd, cl)
slingCurves(sds)
}
