% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\docType{methods}
\name{slingLineages}
\alias{slingLineages}
\alias{slingLineages,SlingshotDataSet-method}
\alias{slingLineages,SingleCellExperiment-method}
\title{Extract the Slingshot lineages}
\usage{
slingLineages(x)

\S4method{slingLineages}{SlingshotDataSet}(x)

\S4method{slingLineages}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{an object containing \code{\link{slingshot}} output.}
}
\value{
the list of lineages, represented by ordered sets of clusters.
}
\description{
Extract lineages (represented by ordered sets of clusters) from
  a \code{SlingshotDataSet}.
}
\section{Methods (by class)}{
\itemize{
\item \code{SlingshotDataSet}: returns the list of lineages, represented by
ordered sets of clusters from a \code{\link{SlingshotDataSet}} object.

\item \code{SingleCellExperiment}: returns the list of lineages, represented by 
ordered sets of clusters from a \code{\link{SingleCellExperiment}} object.
}}

\examples{
data("slingshotExample")
sds <- getLineages(rd, cl)
slingLineages(sds)
}
