% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\name{slingParams}
\alias{slingParams}
\alias{slingParams,PseudotimeOrdering-method}
\alias{slingParams,SingleCellExperiment-method}
\alias{slingParams,SlingshotDataSet-method}
\title{Methods for parameters used by Slingshot}
\usage{
slingParams(x)

\S4method{slingParams}{PseudotimeOrdering}(x)

\S4method{slingParams}{SingleCellExperiment}(x)

\S4method{slingParams}{SlingshotDataSet}(x)
}
\arguments{
\item{x}{an object containing \code{\link{slingshot}} output.}
}
\value{
The list of additional parameters used by Slingshot.
}
\description{
Extracts additional control parameters used by Slingshot in
lineage inference and fitting simultaneous principal curves.
}
\examples{
data("slingshotExample")
rd <- slingshotExample$rd
cl <- slingshotExample$cl
pto <- slingshot(rd, cl, start.clus = '1')
slingParams(pto)
}
