% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\docType{methods}
\name{slingParams}
\alias{slingParams}
\alias{slingParams,SlingshotDataSet-method}
\alias{slingParams,SingleCellExperiment-method}
\title{Methods for parameters used by Slingshot}
\usage{
slingParams(x)

\S4method{slingParams}{SlingshotDataSet}(x)

\S4method{slingParams}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{an object containing \code{\link{slingshot}} output.}
}
\value{
the list of additional parameters used by Slingshot.
}
\description{
Extracts additional control parameters used by Slingshot in 
lineage inference and fitting simultaneous principal curves.
}
\section{Methods (by class)}{
\itemize{
\item \code{SlingshotDataSet}: returns the list of additional parameters used
by Slingshot from a \code{\link{SlingshotDataSet}} object.

\item \code{SingleCellExperiment}: returns the list of additional parameters used
by Slingshot from a \code{\link{SingleCellExperiment}} object.
}}

\examples{
data("slingshotExample")
sds <- slingshot(rd, cl, start.clus = '5')
slingParams(sds)
}
