% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\docType{methods}
\name{slingPseudotime}
\alias{slingPseudotime}
\alias{slingCurveWeights}
\alias{slingPseudotime,SlingshotDataSet-method}
\alias{slingPseudotime,SingleCellExperiment-method}
\alias{slingCurveWeights,SlingshotDataSet-method}
\alias{slingCurveWeights,SingleCellExperiment-method}
\title{Get Slingshot pseudotime values}
\usage{
slingPseudotime(x, ...)

slingCurveWeights(x)

\S4method{slingPseudotime}{SlingshotDataSet}(x, na = TRUE)

\S4method{slingPseudotime}{SingleCellExperiment}(x, na = TRUE)

\S4method{slingCurveWeights}{SlingshotDataSet}(x)

\S4method{slingCurveWeights}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{an object containing \code{\link{slingshot}} output.}

\item{...}{additional parameters to be passed to object-specific methods.}

\item{na}{logical. If \code{TRUE} (default), cells that are not assigned to a
lineage will have a pseudotime value of \code{NA}. Otherwise, their
arclength along each curve will be returned.}
}
\value{
an \code{n} by \code{L} matrix representing each cell's pseudotime
  along each lineage.

an \code{n} by \code{L} matrix of cell weights along
  each lineage.
}
\description{
Extract the matrix of pseudotime values or cells' weights along
  each lineage.
}
\section{Methods (by class)}{
\itemize{
\item \code{SlingshotDataSet}: returns the matrix of pseudotime values from a
\code{\link{SlingshotDataSet}} object.

\item \code{SingleCellExperiment}: returns the matrix of pseudotime values from a
\code{\link{SingleCellExperiment}} object.

\item \code{SlingshotDataSet}: returns the matrix of cell weights along each
lineage from a \code{\link{SlingshotDataSet}} object.

\item \code{SingleCellExperiment}: returns the matrix of cell weights along each
lineage from a \code{\link{SingleCellExperiment}} object.
}}

\examples{
data("slingshotExample")
sds <- slingshot(rd, cl)
slingPseudotime(sds)
data("slingshotExample")
sds <- slingshot(rd, cl)
slingCurveWeights(sds)
}
