% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\docType{methods}
\name{SlingshotDataSet-methods}
\alias{SlingshotDataSet-methods}
\alias{reducedDim,SlingshotDataSet-method}
\alias{clusterLabels,SlingshotDataSet-method}
\alias{connectivity,SlingshotDataSet-method}
\alias{lineages,SlingshotDataSet-method}
\alias{lineageControl,SlingshotDataSet-method}
\alias{curves,SlingshotDataSet-method}
\alias{pseudotime,SlingshotDataSet-method}
\alias{curveWeights,SlingshotDataSet-method}
\alias{curveControl,SlingshotDataSet-method}
\alias{reducedDim<-,SlingshotDataSet-method}
\alias{clusterLabels<-,SlingshotDataSet-method}
\alias{[,SlingshotDataSet-method}
\title{Methods for the \code{SlingshotDataSet} class}
\usage{
\S4method{show}{SlingshotDataSet}(object)

\S4method{reducedDim}{SlingshotDataSet}(x)

\S4method{clusterLabels}{SlingshotDataSet}(x)

\S4method{connectivity}{SlingshotDataSet}(x)

\S4method{lineages}{SlingshotDataSet}(x)

\S4method{lineageControl}{SlingshotDataSet}(x)

\S4method{curves}{SlingshotDataSet}(x)

\S4method{pseudotime}{SlingshotDataSet}(x, na = TRUE)

\S4method{curveWeights}{SlingshotDataSet}(x)

\S4method{curveControl}{SlingshotDataSet}(x)

\S4method{reducedDim}{SlingshotDataSet}(x) <- value

\S4method{clusterLabels}{SlingshotDataSet}(x) <- value

\S4method{[}{SlingshotDataSet}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{na}{logical. If \code{TRUE}, cells that are not assigned to a lineage
will have a pseudotime value of \code{NA}. Otherwise, their arclength along
the curve will be returned.}
}
\description{
Helper functions for accessing and changing elements of a
  \code{SlingshotDataSet} object.
}
