% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\docType{methods}
\name{SlingshotDataSet-methods}
\alias{SlingshotDataSet-methods}
\alias{reducedDim,SlingshotDataSet-method}
\alias{clusterLabels,SlingshotDataSet-method}
\alias{connectivity,SlingshotDataSet-method}
\alias{lineages,SlingshotDataSet-method}
\alias{lineageControl,SlingshotDataSet-method}
\alias{curves,SlingshotDataSet-method}
\alias{curveControl,SlingshotDataSet-method}
\alias{reducedDim<-,SlingshotDataSet-method}
\alias{clusterLabels<-,SlingshotDataSet-method}
\alias{[,SlingshotDataSet,ANY,ANY,ANY-method}
\title{Methods for the \code{SlingshotDataSet} class}
\usage{
\S4method{show}{SlingshotDataSet}(object)

\S4method{reducedDim}{SlingshotDataSet}(x)

\S4method{clusterLabels}{SlingshotDataSet}(x)

\S4method{connectivity}{SlingshotDataSet}(x)

\S4method{lineages}{SlingshotDataSet}(x)

\S4method{lineageControl}{SlingshotDataSet}(x)

\S4method{curves}{SlingshotDataSet}(x)

\S4method{curveControl}{SlingshotDataSet}(x)

\S4method{reducedDim}{SlingshotDataSet}(x) <- value

\S4method{clusterLabels}{SlingshotDataSet}(x) <- value

\S4method{[}{SlingshotDataSet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{a \code{SlingshotDataSet} object.}
}
\description{
Helper functions for accessing and changing elements of a
  \code{SlingshotDataSet} object.
}
\seealso{
\code{\link{SlingshotDataSet}}
}
