% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{SlingshotDataSet}
\alias{SlingshotDataSet}
\title{Extract Slingshot output from SingleCellExperiment}
\usage{
SlingshotDataSet(data, ...)
}
\arguments{
\item{data}{a \code{SingleCellExperiment} object produced by 
\code{slingshot}.}
}
\description{
This is a convenience function to extract \code{slingshot} 
outputs (in the form of a \code{SlingshotDataSet}) from a 
\code{SingleCellExperiment} object produced by \code{slingshot}
}
