% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllHelperFunctions.R
\docType{methods}
\name{SlingshotDataSet}
\alias{SlingshotDataSet}
\alias{SlingshotDataSet,SingleCellExperiment-method}
\title{Extract Slingshot output}
\usage{
SlingshotDataSet(data, ...)

\S4method{SlingshotDataSet}{SingleCellExperiment}(data)
}
\arguments{
\item{data}{an object containing \code{\link{slingshot}} output.}

\item{...}{additional arguments to pass to object-specific methods.}
}
\description{
This is a convenience function to extract a
  \code{SlingshotDataSet} from an object containing \code{\link{slingshot}}
  output.
}
