% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllHelperFunctions.R
\docType{data}
\name{slingshotExample}
\alias{slingshotExample}
\title{Bifurcating lineages data}
\format{
\code{rd} is a matrix of coordinates in two dimensions, representing
  140 cells. \code{cl} is a numeric vector of 140 corresponding cluster
  labels for each cell.
}
\source{
Simulated data provided with the \code{slingshot} package.
}
\usage{
data("slingshotExample")
}
\description{
This simulated dataset contains a low-dimensional representation
  of two bifurcating lineages (\code{rd}) and a vector of cluster labels
  generated by k-means with \code{K = 5} (\code{cl}).
}
\examples{
data("slingshotExample")
rd <- slingshotExample$rd
cl <- slingshotExample$cl
slingshot(rd, cl)
}
\keyword{datasets}
